+++ @utahdts/utah-design-system/react/components/tooltip/Tooltip.jsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from 'react';
-import { usePopper } from 'react-popper';
+import { useFloating, autoUpdate, offset as floatingOffset, shift, flip, arrow } from '@floating-ui/react-dom';
 import { popupPlacement } from '../../enums/popupPlacement';
 import { usePopupDelay } from '../../hooks/usePopupDelay';
-import { useRefAlways } from '../../hooks/useRefAlways';
 import { joinClassNames } from '../../util/joinClassNames';
 
 /** @typedef {import('@utahdts/utah-design-system-header').PopupPlacement} PopupPlacement */
@@ -10,13 +9,14 @@ import { joinClassNames } from '../../util/joinClassNames';
 /**
  * A ToolTip is only in charge of positioning and rendering a tooltip.
  * Pass in a "referenceElement" to have "zero-config" for onMouseEnter and onMouseLeave triggering
- * Pass in a isPopperVisible and setIsPopperVisible to have it be a controlled component
+ * Pass in a isPopupVisible and setIsPopupVisible to have it be a controlled component
  * @param {object} props
  * @param {import('react').ReactNode} props.children The content of the tool tip
  * @param {string} [props.className] CSS class to apply to the popup
  * @param {import('react').MutableRefObject<HTMLDivElement | null>} [props.innerRef] ref of the popup wrapper
- * @param {boolean} [props.isPopperVisible] controlled value for telling if tool tip is visible
- * @param {[number, number]} [props.offset] default offset is [0, 5] (see popper documentation for details)
+ * @param {boolean} [props.isPopupVisible] controlled value for telling if tool tip is visible
+ * @param {number | {mainAxis: number, crossAxis: number, alignmentAxis?: number}} [props.offset] default offset is 5 (see popup documentation
+ * for details)
  * @param {PopupPlacement} [props.placement] where to put the tooltip in reference to the referenceElement
  * @param {HTMLElement | null} props.referenceElement the referenceElement from which the tool tip will toggle (first render will most likely be null)
  * @returns {import('react').JSX.Element}
@@ -25,44 +25,38 @@ export function Tooltip({
   children,
   className,
   innerRef: draftInnerRef,
-  isPopperVisible,
-  offset = [0, 5],
+  isPopupVisible,
+  offset = 5,
   placement = popupPlacement.BOTTOM,
   referenceElement: draftReferenceElement,
 }) {
-  const [isPopperVisibleInternal, setIsPopperVisibleInternal] = useState(false);
-  const [popperElement, setPopperElement] = /** @type {typeof useState<HTMLDivElement | null>} */ (useState)(null);
+  const [isPopupVisibleInternal, setIsPopupVisibleInternal] = useState(false);
+  const [popupElement, setPopupElement] = /** @type {typeof useState<HTMLDivElement | null>} */ (useState)(null);
   const [arrowElement, setArrowElement] = /** @type {typeof useState<HTMLDivElement | null>} */ (useState)(null);
-  const { styles, attributes, update } = usePopper(
-    draftReferenceElement,
-    popperElement,
-    {
-      placement,
-      modifiers: [
-        {
-          name: 'offset',
-          options: { offset },
-        },
-        {
-          name: 'arrow',
-          options: { element: arrowElement },
-        },
-      ],
-    }
-  );
-  const updateRef = useRefAlways(update);
+  const { floatingStyles, middlewareData } = useFloating({
+    elements: {
+      reference: draftReferenceElement,
+      floating: popupElement,
+    },
+    middleware: [
+      floatingOffset(offset),
+      flip(),
+      shift(),
+      arrow({
+        element: arrowElement,
+      }),
+    ],
+    open: !(isPopupVisible ?? isPopupVisibleInternal),
+    placement,
+    whileElementsMounted: autoUpdate,
+  });
 
   const { startNoPopupTimer, startPopupTimer } = usePopupDelay();
-
-  // When the children update recalculate the popper position
-  useEffect(() => {
-    updateRef.current?.();
-  }, [children, updateRef]);
 
   useEffect(
     () => {
       // parent is not controlling visibility, so hookup visibility to the `referenceElement`
-      if (draftReferenceElement && isPopperVisible === undefined) {
+      if (draftReferenceElement && isPopupVisible === undefined) {
         if (draftReferenceElement.onmouseenter) {
           throw new Error('ToolTip: onMouseEnter previously set');
         }
@@ -77,20 +71,18 @@ export function Tooltip({
         }
         draftReferenceElement.onmouseenter = () => (
           startPopupTimer(() => {
-            setIsPopperVisibleInternal(true);
-            updateRef.current?.();
+            setIsPopupVisibleInternal(true);
           })
         );
         // onfocus and onblur don't wait on the popupTimer to popup
         draftReferenceElement.onfocus = () => {
-          setIsPopperVisibleInternal(true);
-          updateRef.current?.();
+          setIsPopupVisibleInternal(true);
         };
         draftReferenceElement.onmouseleave = () => {
           startNoPopupTimer();
-          setIsPopperVisibleInternal(false);
+          setIsPopupVisibleInternal(false);
         };
-        draftReferenceElement.onblur = () => setIsPopperVisibleInternal(false);
+        draftReferenceElement.onblur = () => setIsPopupVisibleInternal(false);
       }
       return (
         () => {
@@ -103,29 +95,29 @@ export function Tooltip({
         }
       );
     },
-    [draftReferenceElement, isPopperVisible, startNoPopupTimer, startPopupTimer, updateRef]
+    [draftReferenceElement, isPopupVisible, startNoPopupTimer, startPopupTimer]
   );
 
   return (
     <div
       ref={(refValue) => {
-        setPopperElement(refValue);
+        setPopupElement(refValue);
         if (draftInnerRef) {
           draftInnerRef.current = refValue;
         }
       }}
-      style={styles.popper}
-      {...attributes.popper}
+      style={floatingStyles}
       className={joinClassNames(
         className,
         'tooltip__wrapper',
-        (!(isPopperVisible ?? isPopperVisibleInternal)) && 'tooltip__wrapper--hidden'
+        (!(isPopupVisible ?? isPopupVisibleInternal)) && 'tooltip__wrapper--hidden'
       )}
       aria-hidden="true"
+      data-popup-placement={middlewareData?.offset?.placement || placement}
     >
       <div className="tooltip__content">
         {children}
-        <div ref={setArrowElement} style={styles.arrow} className="tooltip__arrow" />
+        <div ref={setArrowElement} style={{left: middlewareData.arrow?.x, top: middlewareData.arrow?.y}} className="tooltip__arrow" />
       </div>
     </div>
   );